requirements.txt
requests
argparse
dns.resolver
colorama
random
